+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - BalanceResource: Ignore resources which are in a non-valid replication state
 - sos-report: include backup ship queue
+- sos-report: syslog will only be gathered if messages was not available
 
 ## [1.30.4] - 2025-02-03